refactor: enum -> as const +  폴더 구조 변경(features 폴더로 이동)
