--- src/components/ControllerBar/index.tsx
@@ -1,1 +1,0 @@
-export { default } from "./ControllerBar";
