+++ src/features/controllerBar/components/icons/CheckboxIcon.tsx
