+++ src/features/controllerBar/components/icons/ControlIcon.tsx
