+++ src/features/controllerBar/components/snap/Checkbox.tsx
@@ -1,7 +1,7 @@
 import { ChangeEvent, useState } from "react";
 import { styled } from "styled-components";
 
-import CheckboxIcon from "@components/ControllerBar/icons/CheckboxIcon";
+import CheckboxIcon from "@/features/controllerBar/components/icons/CheckboxIcon";
 
 interface Props {
   value: string;
