+++ src/features/controllerBar/constants/gizmo.ts
@@ -1,0 +1,4 @@
+export const GIZMO_MODE = {
+  GLOBAL: "GLOBAL",
+  LOCAL: "LOCAL",
+} as const;
