+++ src/features/controllerBar/constants/snap.ts
@@ -1,0 +1,5 @@
+export const SNAP_MODE = {
+  GRID: "GRID",
+  ROTATE: "ROTATE",
+  SURFACE: "SURFACE",
+} as const;
