+++ src/features/controllerBar/types/gizmo.ts
@@ -1,0 +1,3 @@
+import { GIZMO_MODE } from "@/features/controllerBar/constants/gizmo";
+
+export type GizmoModeType = (typeof GIZMO_MODE)[keyof typeof GIZMO_MODE];
