+++ src/features/controllerBar/types/snap.ts
@@ -1,0 +1,3 @@
+import { SNAP_MODE } from "@/features/controllerBar/constants/snap";
+
+export type SnapModeType = (typeof SNAP_MODE)[keyof typeof SNAP_MODE];
