+++ src/store/storeContainer.ts
@@ -11,7 +11,7 @@ import renderStore from "./renderStore";
 import assetLibraryStore from "./assetLibraryStore";
 import assetCategoryStore from "./assetCategoryStore";
 import sceneSettingStore from "./sceneSettingStore";
-import controllerBarStore from "./controllerBarStore";
+import controllerBarStore from "@/features/controllerBar/store/controllerBarStore";
 
 const storeContainer = {
   primitiveStore: primitiveStore,
